refactor: restaurant type 템플릿파일로 위치 변경
